Extract DatabaseList
