+++ frontend/src/metabase/common/components/EntityPicker/components/NotebookDataPicker/TableItemPickerResolver.tsx
@@ -1,15 +1,13 @@
-import { useDatabaseListQuery } from "metabase/common/hooks";
-import LoadingAndErrorWrapper from "metabase/components/LoadingAndErrorWrapper";
 import { PERSONAL_COLLECTIONS } from "metabase/entities/collections";
 
-import type { EntityPickerOptions, TisFolder } from "../../types";
+import type { EntityPickerOptions } from "../../types";
 import {
   EntityItemList,
-  ItemList,
   PersonalCollectionsItemList,
   type EntityItemListProps,
 } from "../ItemList";
 
+import { DatabaseList } from "./DatabaseList";
 import type { NotebookDataPickerItem } from "./types";
 
 export const TableItemPickerResolver = ({
@@ -24,7 +22,7 @@ export const TableItemPickerResolver = ({
 }) => {
   if (!query) {
     return (
-      <RootItemList
+      <DatabaseList
         isCurrentLevel={isCurrentLevel}
         isFolder={isFolder}
         selectedItem={selectedItem}
@@ -51,52 +49,6 @@ export const TableItemPickerResolver = ({
       selectedItem={selectedItem}
       isFolder={isFolder}
       isCurrentLevel={isCurrentLevel}
-    />
-  );
-};
-
-interface RootItemListProps {
-  selectedItem: NotebookDataPickerItem | null;
-  isFolder: TisFolder<NotebookDataPickerItem>;
-  isCurrentLevel: boolean;
-  onClick: (val: NotebookDataPickerItem) => void;
-}
-
-const RootItemList = ({
-  isCurrentLevel,
-  isFolder,
-  selectedItem,
-  onClick,
-}: RootItemListProps) => {
-  const {
-    data: databases = [],
-    error,
-    isLoading,
-  } = useDatabaseListQuery({
-    query: { saved: false }, // saved questions are fetched in a separate tab
-  });
-
-  const items = databases.map((database): NotebookDataPickerItem => {
-    return {
-      description: database.description,
-      id: database.id,
-      model: "database",
-      name: database.displayName(),
-    };
-  });
-
-  if (error) {
-    return <LoadingAndErrorWrapper error={error} />;
-  }
-
-  return (
-    <ItemList
-      isCurrentLevel={isCurrentLevel}
-      isFolder={isFolder}
-      isLoading={isLoading}
-      items={items}
-      selectedItem={selectedItem}
-      onClick={onClick}
     />
   );
 };
